+++ example/lib/intro_screen_custom_layout.dart
@@ -50,7 +50,8 @@ class IntroScreenCustomLayoutState extends State<IntroScreenCustomLayout> {
   ButtonStyle myButtonStyle() {
     return ButtonStyle(
       shape: MaterialStateProperty.all<OutlinedBorder>(const StadiumBorder()),
-      backgroundColor: MaterialStateProperty.all<Color>(const Color(0x33ffcc5c)),
+      backgroundColor:
+          MaterialStateProperty.all<Color>(const Color(0x33ffcc5c)),
       overlayColor: MaterialStateProperty.all<Color>(const Color(0x33ffcc5c)),
     );
   }
@@ -58,7 +59,8 @@ class IntroScreenCustomLayoutState extends State<IntroScreenCustomLayout> {
   List<Widget> generateListCustomTabs() {
     return List.generate(
       3,
-      (index) => SizedBox(
+      (index) => Container(
+        color: Colors.black26,
         width: double.infinity,
         height: double.infinity,
         child: ListView(
@@ -75,10 +77,12 @@ class IntroScreenCustomLayoutState extends State<IntroScreenCustomLayout> {
                   color: secondColor,
                 ),
                 onChanged: (String? value) {},
-                items: ["0", "1", "2"].map<DropdownMenuItem<String>>((String value) {
+                items: ["0", "1", "2"]
+                    .map<DropdownMenuItem<String>>((String value) {
                   return DropdownMenuItem<String>(
                     value: value,
-                    child: Text(value, style: TextStyle(color: secondColor, fontSize: 20)),
+                    child: Text(value,
+                        style: TextStyle(color: secondColor, fontSize: 20)),
                   );
                 }).toList(),
               ),