+++ example/lib/intro_screen_default.dart
@@ -20,7 +20,8 @@ class IntroScreenDefaultState extends State<IntroScreenDefault> {
     listContentConfig.add(
       const ContentConfig(
         title: "ERASER",
-        description: "Allow miles wound place the leave had. To sitting subject no improve studied limited",
+        description:
+            "Allow miles wound place the leave had. To sitting subject no improve studied limited",
         pathImage: "images/photo_eraser.png",
         backgroundColor: Color(0xfff5a623),
       ),
@@ -28,7 +29,8 @@ class IntroScreenDefaultState extends State<IntroScreenDefault> {
     listContentConfig.add(
       const ContentConfig(
         title: "PENCIL",
-        description: "Ye indulgence unreserved connection alteration appearance",
+        description:
+            "Ye indulgence unreserved connection alteration appearance",
         pathImage: "images/photo_pencil.png",
         backgroundColor: Color(0xff203152),
       ),